Move controller tests to controllers folder
